+++ src/base/GlobalSetProcessor.php
@@ -84,16 +84,13 @@ class GlobalSetProcessor extends Processor
             $attributeObj[$attribute] = $item->$attribute;
         }
 
+        list ($fieldLayout, $fieldConfigs) = $this->exportFieldLayout($item->getFieldLayout());
         $globalSetObj = array_merge([
             'name' => $item->name,
             'handle' => $item->handle,
-            'fieldLayout' => $this->exportFieldLayout($item->getFieldLayout()),
-            'requiredFields' => $this->exportRequiredFields($item->getFieldLayout()),
+            'fieldLayout' => $fieldLayout,
+            'fieldConfigs' => $fieldConfigs,
         ], $attributeObj);
-
-        if (\count($globalSetObj['requiredFields']) <= 0) {
-            unset($globalSetObj['requiredFields']);
-        }
 
         return $this->stripNulls($globalSetObj);
     }