+++ src/base/TagGroupProcessor.php
@@ -104,16 +104,13 @@ class TagGroupProcessor extends Processor
             $attributeObj[$attribute] = $item->$attribute;
         }
 
+        list ($fieldLayout, $fieldConfigs) = $this->exportFieldLayout($item->getFieldLayout());
         $tagGroupObj = array_merge([
             'name' => $item->name,
             'handle' => $item->handle,
-            'fieldLayout' => $this->exportFieldLayout($item->getFieldLayout()),
-            'requiredFields' => $this->exportRequiredFields($item->getFieldLayout()),
+            'fieldLayout' => $fieldLayout,
+            'fieldConfigs' => $fieldConfigs,
         ], $attributeObj);
-
-        if (\count($tagGroupObj['requiredFields']) <= 0) {
-            unset($tagGroupObj['requiredFields']);
-        }
 
         return $this->stripNulls($tagGroupObj);
     }